+++ CHANGES.md
@@ -4,6 +4,9 @@
 ### Changes
 * Raise minimum golang requirement to 1.17
 
+### Features
+* Add option to pass extra environment variables to the execution via `-extra-env`
+
 ## v1.7.0 (2023-05-23)
 ### Changes
 * No changes, just a re-tag of v1.7.0-rc0